requirements.txt
Flask==2.3.0
Flask-CORS==4.0.0
Flask-SocketIO==5.3.4
Flask-Login==0.6.3
python-socketio==5.9.0
pycryptodome==3.18.0
requests==2.31.0
eventlet==0.33.3
cryptography==41.0.0
bcrypt==4.1.2
qrcode==7.4.2
Pillow==10.4.0
python-dotenv==1.0.0
